rename package
(cherry picked from commit 4f78efd)
